+++ core/src/com/stone/core/msg/ProtobufMessageFactory.java
@@ -11,7 +11,7 @@ import com.stone.core.codec.IMessageFactory;
 public class ProtobufMessageFactory implements IMessageFactory {
 
 	@Override
-	public IProtobufMessage createMessage(short type) {
+	public IProtobufMessage createMessage(int type) {
 		ProtobufMessage message = new ProtobufMessage(type);
 		return message;
 	}
